switching to eventLog name
